requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
openpyxl
prisma
adaptivetesting
python-dotenv
